# openapi.with-code-samples.yml
openapi: 3.0.3
info:
  title: OpenAI API (Unofficial Swagger Template)
  version: "2026-02-16"
  description: >
    Swagger-ready OpenAPI spec for common OpenAI endpoints.
    This is a pragmatic template (not an official full schema). Unknown fields are allowed.
servers:
  - url: https://api.openai.com/v1
...
security:
  - bearerAuth: []
paths:
  /models:
    get:
      tags: [Models]
      operationId: listModels
      summary: List models
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsList"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /models/{model}:
    get:
      tags: [Models]
      operationId: retrieveModel
      summary: Retrieve a model
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /files:
    get:
      tags: [Files]
      operationId: listFiles
      summary: List files
      parameters:
        - name: after
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 10000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileList"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Files]
      operationId: createFile
      summary: Upload a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileCreateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileObject"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /files/{file_id}:
    get:
      tags: [Files]
      operationId: retrieveFile
      summary: Retrieve a file
      parameters:
        - name: file_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileObject"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Files]
      operationId: deleteFile
      summary: Delete a file
      parameters:
        - name: file_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteObject"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /files/{file_id}/content:
    get:
      tags: [Files]
      operationId: retrieveFileContent
      summary: Retrieve file content
      parameters:
        - name: file_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Raw file content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "API key"
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string }
            param: { type: string, nullable: true }
            code: { type: string, nullable: true }
          required: [message, type]
      required: [error]
    DeleteObject:
      type: object
      properties:
        id: { type: string }
        object: { type: string }
        deleted: { type: boolean }
      required: [id, deleted, object]
      additionalProperties: true
    Model:
      type: object
      properties:
        id: { type: string }
        object: { type: string, example: "model" }
        created: { type: integer, nullable: true }
        owned_by: { type: string, nullable: true }
      required: [id, object]
      additionalProperties: true
    ModelsList:
      type: object
      properties:
        object: { type: string, example: "list" }
        data:
          type: array
          items:
            $ref: "#/components/schemas/Model"
      required: [object, data]
      additionalProperties: true
...
    FileObject:
      type: object
      properties:
        id: { type: string }
        object: { type: string, example: "file" }
        bytes: { type: integer, nullable: true }
        created_at: { type: integer, nullable: true }
        filename: { type: string, nullable: true }
        purpose: { type: string, nullable: true }
      required: [id, object]
      additionalProperties: true
    FileList:
      type: object
      properties:
        object: { type: string, example: "list" }
        data:
          type: array
          items:
            $ref: "#/components/schemas/FileObject"
      required: [object, data]
      additionalProperties: true
    FileCreateRequest:
      type: object
      required: [file, purpose]
      properties:
        file:
          type: string
          format: binary
        purpose:
          type: string
      additionalProperties: true
...